requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
alembic==1.14.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   starlette
    #   watchfiles
async-timeout==5.0.1
    # via asyncpg
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==4.2.1
    # via -r requirements.in
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.0
    # via python-jose
email-validator==2.2.0
    # via pydantic
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.8
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
mako==1.3.9
    # via alembic
markupsafe==3.0.2
    # via mako
passlib==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic[email]==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via uvicorn
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via uvicorn
rsa==4.9
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy[asyncio]==2.0.38
    # via
    #   -r requirements.in
    #   alembic
starlette==0.45.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
uvicorn[standard]==0.34.0
    # via -r requirements.in
watchfiles==1.0.4
    # via uvicorn
websockets==14.2
    # via uvicorn
